Finish hooking up represenetatives API call
