+++ src/main/java/co/willsalz/ttyl/resources/v1/RepresentativeResource.java
@@ -1,16 +1,16 @@
 package co.willsalz.ttyl.resources.v1;
 
-import co.willsalz.ttyl.entities.RepresentativeInfo;
+import co.willsalz.ttyl.entities.Representatives;
 import co.willsalz.ttyl.gateways.RepresentativeLookupGateway;
 import org.hibernate.validator.constraints.NotEmpty;
 
-import javax.ws.rs.*;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
-import java.util.List;
 
-/**
- * Created by ahb5 on 1/27/17.
- */
 @Path("v1/representatives")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
@@ -22,7 +22,7 @@ public class RepresentativeResource {
     }
 
     @GET
-    public List<RepresentativeInfo> getRepresentatives(@QueryParam("zip") @NotEmpty String zip) {
+    public Representatives getRepresentatives(@QueryParam("zip") @NotEmpty String zip) {
         return representativeLookupGateway.getRepresentativesByZip(zip);
     }
 }
